Re-organize web client source files
