--- web-client/src/bls/mod.rs
@@ -1,3 +1,0 @@
-pub mod key_pair;
-pub mod public_key;
-pub mod secret_key;
