+++ web-client/src/client/mod.rs
@@ -1,0 +1,5 @@
+pub mod account;
+pub mod block;
+pub mod client;
+pub mod client_configuration;
+pub mod peer_info;
