+++ web-client/src/client/peer_info.rs
