+++ web-client/src/lib.rs
@@ -1,28 +1,9 @@
 extern crate alloc; // Required for wasm-bindgen-derive
 
-#[cfg(feature = "client")]
-mod account;
 mod address;
 #[cfg(feature = "client")]
-mod block;
-#[cfg(feature = "primitives")]
-mod bls;
-#[cfg(feature = "client")]
 mod client;
-mod client_configuration;
 #[cfg(feature = "primitives")]
-mod key_pair;
-#[cfg(feature = "client")]
-mod peer_info;
-#[cfg(feature = "primitives")]
-mod private_key;
-#[cfg(feature = "primitives")]
-mod public_key;
-#[cfg(feature = "primitives")]
-mod signature;
-#[cfg(feature = "primitives")]
-mod signature_proof;
+mod primitives;
 mod transaction;
-#[cfg(feature = "primitives")]
-mod transaction_builder;
 mod utils;
