+++ web-client/src/primitives/public_key.rs
@@ -3,7 +3,9 @@ use std::str::FromStr;
 use nimiq_serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::*;
 
-use crate::{address::Address, private_key::PrivateKey, signature::Signature};
+use crate::{
+    address::Address, primitives::private_key::PrivateKey, primitives::signature::Signature,
+};
 
 /// The non-secret (public) part of an asymmetric key pair that is typically used to digitally verify or encrypt data.
 #[wasm_bindgen]
