+++ web-client/src/primitives/signature_proof.rs
@@ -1,7 +1,9 @@
 use nimiq_serde::Serialize;
 use wasm_bindgen::prelude::*;
 
-use crate::{address::Address, public_key::PublicKey, signature::Signature};
+use crate::{
+    address::Address, primitives::public_key::PublicKey, primitives::signature::Signature,
+};
 
 /// A signature proof represents a signature together with its public key and the public key's merkle path.
 /// It is used as the proof for transactions.
